Correções do merge

+++ src/app/PrimaryComparator.java
@@ -6,7 +6,7 @@ public class PrimaryComparator implements Comparator<Musica> {
     
     public int compare(Musica m1, Musica m2) {
 
-        return Integer.compare(m1.getIndex(), m2.getIndex());
+        return Long.compare(m1.getIndex(), m2.getIndex());
 
     }
